Refactor app structure into commands/, services/, utils/ directories
Split the flat src/ structure into organized directories:
- commands/ — individual command handlers (download, checkDiff, apply, common)
- services/ — vault interaction layer (VaultService, VaultNormalizer)
- utils/ — generic utilities (diff, process, file)

Removed monolithic Helpers.ts and Logic.ts in favor of focused,
single-responsibility modules. Updated test imports accordingly.
