+++ src/services/VaultNormalizer.ts
